+++ packages/components/src/components/form/form-styles.ts
@@ -1,0 +1,138 @@
+import type { Styles } from '../../utils';
+import {
+  addImportantToRule,
+  BreakpointCustomizable,
+  buildResponsiveStyles,
+  colorDarken,
+  getFormTextHiddenJssStyle,
+  getInset,
+  getThemedColors,
+  getThemedFormStateColors,
+  getTransition,
+  isDark,
+  pxToRemWithUnit,
+} from '../../utils';
+import { color, font } from '@porsche-design-system/utilities';
+import { FormState, Theme } from '../../types';
+import { JssStyle } from 'jss';
+
+export const INPUT_HEIGHT = 48;
+export const isVisibleState = (state: FormState): boolean => state === 'success' || state === 'error';
+
+export type ChildSelector = 'input' | 'select' | 'textarea';
+
+export const getBaseChildStyles = (
+  child: ChildSelector,
+  state: FormState,
+  theme: Theme,
+  additionalDefaultJssStyle?: JssStyle
+): Styles => {
+  const { baseColor, backgroundColor, contrastHighColor, contrastMediumColor } = getThemedColors(theme);
+  const { stateColor, stateHoverColor } = getThemedFormStateColors(theme, state);
+  const hasVisibleState = isVisibleState(state);
+
+  const { disabled } = color.state; // 🤷 no theming here
+// TODO: Add readonly color to utilities package
+  const readonly = '#ebebeb'; // 🤷
+
+  return {
+    [`::slotted(${child})`]: {
+      display: 'block',
+      position: 'relative',
+      ...getInset(),
+      width: '100%',
+      ...(child !== 'textarea' && { height: pxToRemWithUnit(INPUT_HEIGHT) }),
+      margin: 0,
+      outline: '1px solid transparent',
+      outlineOffset: '2px',
+      appearance: 'none',
+      WebkitAppearance: 'none', // iOS safari
+      boxSizing: 'border-box',
+      border: hasVisibleState ? `2px solid ${stateColor}` : `1px solid ${contrastMediumColor}`,
+      borderRadius: 0,
+      backgroundColor,
+      opacity: 1,
+      fontFamily: font.family,
+      fontWeight: font.weight.regular,
+      ...font.size.small,
+      textIndent: 0,
+      color: baseColor,
+      transition: ['color', 'border-color', 'background-color'].map(getTransition).join(','),
+      ...additionalDefaultJssStyle,
+    },
+    [`::slotted(${child}:hover)`]: {
+      borderColor: stateHoverColor || (isDark(theme) ? contrastHighColor : baseColor),
+    },
+    [`::slotted(${child}:focus)`]: {
+      outlineColor: stateColor || contrastMediumColor,
+    },
+    [`::slotted(${child}:disabled)`]: {
+      cursor: 'not-allowed',
+      color: disabled, // 🤷
+      borderColor: disabled,
+      WebkitTextFillColor: disabled, // fix placeholder color bug in Safari
+    },
+    ...(child !== 'select' && {
+      [`::slotted(${child}[readonly])`]: {
+        borderColor: readonly,
+        backgroundColor: readonly,
+      },
+      [`::slotted(${child}[readonly]:focus)`]: {
+        outlineColor: 'transparent',
+      },
+      [`::slotted(${child}[readonly]:not(:disabled))`]: {
+        color: contrastMediumColor,
+      },
+    }),
+  };
+};
+
+export const getLabelStyles = (
+  child: ChildSelector,
+  hideLabel: BreakpointCustomizable<boolean>,
+  state: FormState,
+  theme: Theme,
+  additionalRefForInputHover?: string
+): Styles => {
+  const { baseColor, contrastMediumColor, disabledColor } = getThemedColors(theme);
+  const hasVisibleState = isVisibleState(state);
+
+  return {
+    label: {
+      display: 'block',
+      position: 'relative', // for unit and counter
+      '&--disabled': {
+        '& .label__text': {
+          color: disabledColor,
+        },
+        ...(additionalRefForInputHover && {
+          [`& ${additionalRefForInputHover}`]: {
+            color: disabledColor,
+            cursor: 'not-allowed',
+          },
+        }),
+      },
+      '&__text': {
+        ...buildResponsiveStyles(hideLabel, getFormTextHiddenJssStyle),
+        display: 'block',
+        width: 'fit-content',
+        transition: getTransition('color'),
+        '&+&--description': {
+          marginTop: pxToRemWithUnit(-4),
+          paddingBottom: pxToRemWithUnit(8),
+        },
+        '&--description': {
+          color: contrastMediumColor,
+        },
+      },
+    },
+    ['label__text' + (additionalRefForInputHover ? `,${additionalRefForInputHover}` : '')]: {
+      '&:hover': {
+        [`&~::slotted(${child}:not(:disabled):not([readonly]))` +
+        (hasVisibleState ? `, ::slotted(${child}:hover:not(:disabled):not([readonly]))` : '')]: {
+          borderColor: addImportantToRule(hasVisibleState ? colorDarken.notification[state] : baseColor),
+        },
+      },
+    },
+  };
+};