+++ packages/components/src/components/form/text-field-wrapper/text-field-wrapper.tsx
@@ -17,7 +17,14 @@ import type { BreakpointCustomizable, FormState } from '../../../types';
 import { getComponentCss, getSlottedCss } from './text-field-wrapper-styles';
 import { StateMessage } from '../../common/state-message';
 import type { TextFieldWrapperUnitPosition } from './text-field-wrapper-utils';
-import { setInputUnitStyles, throwIfUnitLengthExceeded } from './text-field-wrapper-utils';
+import {
+  addInputEventListener,
+  hasCounterAndIsTypeText,
+  hasUnitAndIsTypeNumber,
+  setCounterInnerHtml,
+  setInputStyles,
+  throwIfUnitLengthExceeded,
+} from './text-field-wrapper-utils';
 
 @Component({
   tag: 'p-text-field-wrapper',
@@ -50,8 +57,10 @@ export class TextFieldWrapper {
   @State() private showPassword = false;
 
   private input: HTMLInputElement;
-  private unitElement: HTMLElement;
+  private unitOrCounterElement: HTMLElement;
   private isPassword: boolean;
+  private hasCounter: boolean;
+  private hasUnit: boolean;
 
   public connectedCallback(): void {
     attachSlottedCss(this.host, getSlottedCss);
@@ -67,6 +76,15 @@ export class TextFieldWrapper {
     );
     this.observeAttributes(); // once initially
     this.isPassword = this.input.type === 'password';
+    this.hasCounter = hasCounterAndIsTypeText(this.input);
+    this.hasUnit = hasUnitAndIsTypeNumber(this.input, this.unit);
+  }
+
+  public componentDidLoad(): void {
+    if (this.hasCounter) {
+      addInputEventListener(this.input, this.unitOrCounterElement, this.setInputStyles);
+      setCounterInnerHtml(this.input, this.unitOrCounterElement); // initial value
+    }
   }
 
   public componentWillRender(): void {
@@ -76,15 +94,15 @@ export class TextFieldWrapper {
       getComponentCss,
       this.hideLabel,
       this.state,
-      this.unit,
-      this.unitPosition,
+      this.hasUnit || this.hasCounter,
+      this.hasCounter ? 'suffix' : this.unitPosition,
       this.isPassword
     );
   }
 
   public componentDidRender(): void {
-    // needs to happen after render in order to have unitElement defined
-    setInputUnitStyles(this.input, this.unit, this.unitElement?.offsetWidth, this.unitPosition, this.state);
+    // needs to happen after render in order to have unitOrCounterElement defined
+    this.setInputStyles();
 
     /*
      * This is a workaround to improve accessibility because the input and the label/description/message text are placed in different DOM.
@@ -109,13 +127,11 @@ export class TextFieldWrapper {
       ['label--disabled']: disabled,
     };
 
-    const unitClasses = {
-      ['unit']: true,
-      ['unit--disabled']: disabled,
+    const labelProps = {
+      tag: 'span',
+      color: 'inherit',
+      onClick: this.onLabelClick,
     };
-
-    const textProps = { tag: 'span', color: 'inherit' };
-    const labelProps = { ...textProps, onClick: this.onLabelClick };
 
     const PrefixedTagNames = getPrefixedTagNames(this.host);
 
@@ -134,14 +150,12 @@ export class TextFieldWrapper {
                 {this.description || <slot name="description" />}
               </PrefixedTagNames.pText>
             )}
-            {type === 'number' && this.unit && (
+            {(this.hasUnit || this.hasCounter) && (
               <PrefixedTagNames.pText
-                class={unitClasses}
-                tag="span"
-                color="inherit"
-                ref={(el) => (this.unitElement = el)}
+                class="unit"
+                {...labelProps}
+                ref={(el) => (this.unitOrCounterElement = el)}
                 aria-hidden="true"
-                onClick={this.onLabelClick}
               >
                 {this.unit}
               </PrefixedTagNames.pText>
@@ -199,5 +213,9 @@ export class TextFieldWrapper {
 
   private observeAttributes = (): void => {
     observeAttributes(this.input, ['disabled', 'readonly', 'required'], () => forceUpdate(this.host));
+  };
+
+  private setInputStyles = (): void => {
+    setInputStyles(this.input, this.unitOrCounterElement, this.hasCounter ? 'suffix' : this.unitPosition, this.state);
   };
 }