+++ packages/components/src/utils/theme.spec.ts
@@ -1,4 +1,4 @@
-import { isDark, getThemeDarkAttribute, isLightElectric } from './theme';
+import { isDark, getThemeDarkAttribute, isLightElectric, isDarkElectric } from './theme';
 
 describe('isDark()', () => {
   it('should return true for "dark"', () => {
@@ -25,6 +25,20 @@ describe('isLightElectric()', () => {
 
   it('should return false for everything else', () => {
     expect(isLightElectric('unknown' as any)).toBe(false);
+  });
+});
+
+describe('isDarkElectric()', () => {
+  it('should return true for "dark-electric"', () => {
+    expect(isDarkElectric('dark-electric')).toBe(true);
+  });
+
+  it('should return false for "light"', () => {
+    expect(isDarkElectric('light')).toBe(false);
+  });
+
+  it('should return false for everything else', () => {
+    expect(isDarkElectric('unknown' as any)).toBe(false);
   });
 });